Restructure and refactor
